+++ pageComponents/folderPageComponents/Card.tsx
