+++ pages/_app.tsx
@@ -1,5 +1,5 @@
 import type { AppProps } from "next/app";
-import "@/styles/landing.css";
+import "@/styles/global.css";
 import Head from "next/head";
 export default function App({ Component, pageProps }: AppProps) {
   return (
